fpstudy: Add script to check whether python3-venv package exists
Check whether the python3-venv package is already installed. Suggest a
means of installing it if it does not exist.

BUG=b:357664111
TEST=./python-venv-setup.sh

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5762534
Commit-Queue: Firas Sammoura <fsammoura@google.com>
Tested-by: Firas Sammoura <fsammoura@google.com>
Commit-Queue: ChromeOS Auto Runner <chromeos-auto-runner@chromeos-bot.iam.gserviceaccount.com>

+++ biod/study/python-venv-setup.sh
@@ -24,6 +24,15 @@ main() {
         loc="${script_dir}/.venv"
     fi
 
+    # Check whether python3-venv package exists.
+    if ! python3 -c 'import venv; import ensurepip' &>/dev/null; then
+        echo "Error - Python3 module venv or ensurepip is missing"
+        echo
+        echo "On Debian, you could do the following:"
+        echo "sudo apt install -y python3-venv"
+        exit 1
+    fi
+
     # Create new python3 virtual environment.
     rm -rf "${loc}"
     if ! python3 -m venv "${loc}"; then
